requirements.txt
aiogram==3.13.1
aiohttp==3.9.5
pandas==2.2.2
python-dotenv==1.0.1